+++ CHANGELOG.md
@@ -54,6 +54,7 @@
   * `Project` searches for issues in a specific project.
   * `Summary` searches for issues whose summary matches a search string.
   * `SubtaskOf` searches for issues that are subtasks of a parent issue.
+* Function `Protect-YTPath` for creating a URL-safe paths to a YouTrack API resource using values that come from users.
 
 ### Changed